Router: Router implementation with GoRouter
